+++ src/encoded/types/submitted_file.py
@@ -2,6 +2,8 @@ from snovault import abstract_collection, load_schema
 
 from .acl import SUBMISSION_CENTER_MEMBER_CREATE_ACL
 from .file import File
+from .submitted_item import SubmittedItem
+
 
 
 @abstract_collection(
@@ -12,7 +14,7 @@ from .file import File
         "title": "SMaHT Submitted Files",
         "description": "Listing of SMaHT Submitted Files",
     })
-class SubmittedFile(File):
+class SubmittedFile(SubmittedItem, File):
     item_type = "submitted_file"
     base_types = ["SubmittedFile"] + File.base_types
     schema = load_schema("encoded:schemas/submitted_file.json")